+++ NEWS.md
@@ -1,5 +1,8 @@
 # dtplyr (development version)
 
+ * The `.cols` argument of `across` is evaluated in the environment from which 
+   `across` was called
+   
 * `if_any()` and `if_all()` now work without specifying the `.fns` argument (@mgirlich, #325).
 
 # dtplyr 1.2.1